added missing measurement quantities current, voltage and power_factor; added SpeedwireProtocol::getTag0() method

+++ include/SpeedwireProtocol.hpp
@@ -33,7 +33,7 @@ public:
 
     bool checkHeader(void);
     uint32_t getSignature(void);
-    uint16_t getTag0(void);
+    uint32_t getTag0(void);
     uint32_t getGroup(void);
     uint16_t getLength(void);
     uint16_t getNetworkVersion(void);
